pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "transaction-mapping"
version = "0.1.0"
description = "CSV mappers for normalising crypto exchange exports"
authors = [{ name = "Max Williams" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyYAML",
]

[project.urls]
Homepage = "https://github.com/maxwilliams94/TCrypto"

[project.scripts]
transaction-mapper = "src.converter.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
